long function to merge reads into sequences
This function does a lot, so here goes…

The basic input is a list of fwd and reverse sequencing reads.

The function then aligns these reads and calls a consensus sequence.

Optionally, one can also pass an amino acid reference sequence, in
which case we ATTEMPT to correct frameshift mutations in the reads, and
then we do a translation alignment. This will (of course) only work if
you have a relatively similar reference sequence. But in the case of
sanger reads, this is almost always the case.

We return a lot of information about the merging process, including the
consensus sequence, an alignment of reads to the consensus, a
dendrogram of the reads, info about frameshift corrections, differences
between each read and the sequence, and a distance matrix between
reads. This is probably overkill, but I can imagine each of these
things being useful in different situations.