fix a bug in trimming
this bug occurred in the very rare case that one single base exactly
matches the cutoff level.

The issue was that I was erroneously resetting the trim_start when this
happened, because I was using <, rather than <=.

Also fixed so that when no bases hit the cutoff, we return trim_finish
= 0, meaning that the sequenced gets trimmed down to 0 bases, rather
than one.